code refactored for better readability. Used more reasonable naming convention. Also, removed preview_ready event
