+++ src-tauri/src/recorder/encoder/config.rs
