+++ src-tauri/src/recorder/encoder/core/utils.rs
